LayoutTests/platform/ios-sim/fast/dom/title-directionality-expected.txt


<html><title>foo</title></html> should have title direction "ltr". PASS
<html><title dir=rtl>foo</title></html> should have title direction "rtl". FAIL (got: "ltr")
<html dir=rtl><title>foo</title></html> should have title direction "rtl". FAIL (got: "ltr")
<html dir=rtl><title dir=ltr>foo</title></html> should have title direction "ltr". PASS
